Add namedtuple to groupAnagrams_test.py
This is an example on how to use namedtuples to increase
readability of code.

When the values of the tuple are long, is easy to loose track to
what field they map to.

With namedtuples, we can refer to the fields by name using the dot
notation and we no longer need to remember what index maps to
what field, increasing readability and maintainability.